Dio Interceptors change

+++ lib/screens/login/token_interceptor.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:hive/hive.dart';
 import 'package:power_lift/data/services/app_service.dart';
 import 'package:power_lift/main.dart';
-import 'package:power_lift/screens/login/state/auth_bloc.dart';
+import 'package:power_lift/navigation/app_router.dart';
 
 class TokenInterceptor extends Interceptor {
   final Box storage = Hive.box('appStorage');
@@ -18,17 +18,13 @@ class TokenInterceptor extends Interceptor {
   @override
   void onError(DioException err, ErrorInterceptorHandler handler) async {
     super.onError(err, handler);
-    kLogger.e('Error. Sign in again.');
-    await storage.delete('token');
-    await storage.delete('username');
-    getIt.get<AuthBloc>().add(const AuthEvent.logOut());
-    // if (err.response?.statusCode == 401) {
-    //   kLogger.e('Token expired. Log in again.');
-    //   await storage.delete('token');
-    //   await storage.delete('username');
-    //   getIt.get<AuthBloc>().add(const AuthEvent.logOut());
-    // } else {
-    //   getIt.get<AuthBloc>().add(const AuthEvent.logOut());
-    // }
+    if (err.response?.statusCode == 401) {
+      kLogger.e('Error. Sign in again.');
+      await storage.delete('token');
+      await storage.delete('username');
+      getIt<AppRouter>().replaceAll([
+        const GetStartedRoute(),
+      ]);
+    }
   }
 }
